in test_scalar, should now use size_t instead of int

+++ tests/examples/scalar/test_scalar.cpp
@@ -25,8 +25,8 @@ TEST(ConvergenceTest, ScalarSDC)
     
   const complex<double> lambda = complex<double>(-1.0, 1.0);
   const double Tend = 4.0;
-  const vector<int> nsteps = { 2, 5, 10, 15, 20 };
-  int nsteps_l = nsteps.size();
+  const vector<size_t> nsteps = { 2, 5, 10, 15, 20 };
+  size_t nsteps_l = nsteps.size();
 
   vector<double> err(nsteps.size());
   vector<double> convrate(nsteps.size()-1);
@@ -36,20 +36,20 @@ TEST(ConvergenceTest, ScalarSDC)
     
   // Test converge up to 6 nodes: For more nodes, errors are of the order of machine 
   // precision and convergence can no longer be monitored  
-  for ( int nnodes = 2; nnodes<=6; ++nnodes)
+  for ( size_t nnodes = 2; nnodes<=6; ++nnodes)
   {
     // Expect convergence rate of 2*nodes-2 from collocation formula,
     // doing an identical number of iteration should suffice to reach this as each
     // iteration should increase order by one
-    int niters = 2*nnodes-2;
+    size_t niters = 2*nnodes-2;
   
-    for ( int i = 0; i<=nsteps_l-1; ++i)
+    for ( size_t i = 0; i<=nsteps_l-1; ++i)
     {
       dt = Tend/double(nsteps[i]);
       err[i] = run_scalar_sdc(nsteps[i], dt, nnodes, niters, lambda);
     }
   
-    for (int i = 0; i<=nsteps_l-2; ++i)
+    for ( size_t i = 0; i<=nsteps_l-2; ++i)
     {
       convrate[i] = log10(err[i+1]/err[i])/log10(double(nsteps[i])/double(nsteps[i+1]));
       // The expected convergence rate for Lobatto nodes is 2*nnodes-2, but because
